Add secure setting for system key in 5.x (#1889)
This is a backport of #1831 that keeps backwards compatibility with the system key file as
the key for encryption of sensitive data in watches. The secure setting will be preferred
over the system_key file and deprecation logging has been added to ensure that the user is made
aware of the need to use the new secure setting for 6.0.